test(compat): split compatibility smoke fixtures
Split the generated TestKit compatibility suite by fixture family so static analysis can keep enforcing size and complexity limits without suppressing the tests.

Keep coverage for Kotlin JVM, KMP target filtering, Android AGP 8/9 routing, Compose KMP, and Compose Desktop generated consumers. The Compose Desktop rows continue to execute createReleaseDistributable instead of package/install tasks to avoid host installer side effects.

Wrap the JVM compatibility warning line to keep root Detekt clean.
